# openapi/server-2.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "server-2",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/auth/steam/login": {
      "get": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "redirectTo",
            "required": false
          }
        ],
        "responses": {
          "302": {
            "description": "Default Response"
          }
        }
      }
    },
    "/auth/steam/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "openid.claimed_id",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "openid.mode",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "redirectTo",
            "required": false
          }
        ],
        "responses": {
          "302": {
            "description": "Default Response"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}